refactor(loading): sfc to tsx
